refactor for ST new regex engine
ST3's new regex engine doesn't work with look-behinds, so refactor the
syntax definition to not use them.
it means the file should be parsed more efficiently now.
side-effect: we have to manually define some xml scopes, so if they
change in the default package, they might not match what we define here