Introduce a custom build command linqpad_exec
This introduces a copy of Default/exec.py that will be tailored to
suite the needs of `lprun` better.

Currently this base version is a version of that file with only the
relative command names modified to provide it a different name than
the default version but in all other respects it behaves the same.

This will be the base of the upcoming enhancements, and having a base
version will make in-progress diff operations easier to visualize.

As a part of this change, the sublime-build file has been updated to
support the new command, including the integration needed to allow the
custom command to cancel the build while it's executing, although this
only seems to work in the more recent dev releases at this point and
I'm unsure as to why.