Implement error message rewriting in linqpad_exec
This modifies the newly added build system target command so that it
buffers input coming from the executed process and only passes it
through to the output panel full lines at a time.

This allows us to pre-apply the file regex so that we can detect any
error messages and rewrite the line number that's being generated so
that it points at the proper place in the file.

This is currently using a hard coded error message offset pending the
implementation of the code that would parse the script before passing
it to lprun in order to determine what the correct value would be.

Also currently the output for stdout and stderr are mingled in the
buffer, even though lprun only generates error messages to stderr.

This imposes some overhead on checking the script output for things
that should not be matching, but for the time being that's good enough
until everything else is working as expected.