Dynamically determine error line offset
This includes into the command the ability to examine the linq script
that is about to be executed in order to determine what offset to
apply to any error messages so that instead of rewriting them with a
known offset it instead uses the one that LINQPad itself would see.

Currently this assumes that what you're building is the currently
active file, which is not necessarily true.

A more robust implementation would instead examine the command line
to be executed in order to determine the correct file (for example
to allow key bindings that execute a predefined file or some such).

Even better would be dynamically performing the lookup when the first
error for a file is found, since it is possible that the command may
reference a script other than the one directly being run, which would
otherwise end up with potentially incorrect line mappings.