Commence command rewriting refactoring
In order to be more robust w/regards to only buffering and rewriting
the stderr stream from the executed command as well as delaying the
operation to determine the script offset until an error is mentioned,
the code for performing these tasks is being pushed up into the
AsyncProcess class.

Since this is the ultimate source of the data it automatically has a
better handle on where the data is coming from. Additionally since the
reading of this data occurs in a thread already, we can safely delay
scanning the script until the first message for it is seen without
having to create another thread first.