Implement stderr buffering within AsyncProcess
This modifies the threads that spin off from AsyncProcess so that the
one that reads from stderr (where lprun errors appear) will
automatically buffer things and only let whole lines through.

Previously the new exec command was doing this itself, which made things
a little more complicated than they technically needed to be.

Currently that command is still in the exec command, which is
temporarily buffering the output only to discover that it always has a
full line of data.