Push the rewriting code into AsyncProcess
This completes the first part of the refactoring of this code by
putting the responsibility of rewriting error messages into the
AsyncProcess class instead of in the command class.

This leaves the command class as mostly stock; in particular it no
longer needs to buffer its own data and can just display as it always
did.

The primary difference between our exec and the custom one now is that
apart from passing more arguments to AsyncProcess, we also don't do
the newline normalization and instead let the AsyncProcess do it for
us, since we need to ensure that we only operate on whole lines while
doing rewriting.